Moved `crate::base62` to root of crate
